layering io.questdb.std.str.{Direct}Utf8Sequence on top of io.questdb.std.bytes.{Direct}ByteSequence

+++ core/src/main/java/io/questdb/std/str/Utf8s.java
@@ -14,7 +14,7 @@ public interface Utf8s {
             return false;
         }
 
-        for (int index = 0, n = a.size(); index < n; index++) {
+        for (long index = 0, n = a.size(); index < n; index++) {
             if (a.byteAt(index) != b.byteAt(index)) {
                 return false;
             }
